requirements.txt
python-dotenv >= 0.15.0
python-telegram-bot >= 13.1
paralleldots >= 3.2.14